Update docs and matcher refactor
