+++ docs/library/matcher/ciratefi.rst
@@ -1,0 +1,10 @@
+:mod:`matcher.ciratefi` --- RST-BC Invariant template match algorithm
+====================================================================
+
+The :mod:`matcher.ciratefi` module
+
+.. versionadded:: 0.1.0
+
+.. automodule:: autocnet.matcher.ciratefi
+   :synopsis:
+   :members:
