+++ docs/library/matcher/index.rst
@@ -7,4 +7,4 @@
    matcher
    outlier_detector
    subpixel
-   
+   ciratefi
